Выполнил ДЗ 5 модуль
